requirements.txt
nltk==3.6.3
sentence-transformers==2.1.0
faiss-cpu==1.7.0
transformers==4.10.0
torch==1.9.0
scikit-learn==0.24.2
numpy==1.21.0     
faiss-cpu==1.8.0
sentence-transformers==3.0.1
nltk==3.8.1